requirements.txt
# requirements.txt
Django==4.2.7
djangorestframework==3.14.0
razorpay==1.4.1
google-generativeai==0.3.2
django-cors-headers==4.3.1
python-decouple==3.8
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
django-extensions==3.2.3
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.5
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0